requirements.txt
streamlit
azure-cognitiveservices-speech
pydub
python-docx
python-dotenv
azure-storage-blob
openai==0.28
azure-identity
